Remove trace, move winlogproc to etc
