+++ etc/winlogproc/cmd/main.go
