+++ etc/winlogproc/collect.go
