+++ etc/winlogproc/collect_other.go
